requirements.txt
schemathesis==3.*
pytest==8.*
requests>=2.32

python-dotenv>=1.0
